btrfs: add ioctl to export size of global metadata reservation
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Jeff Mahoney <jeffm@suse.com>
commit 01e219e8069516cdb98594d417b8bb8d906ed30d

btrfs filesystem df output will show the size of the metadata space
and how much of it is used, and the user assumes that the difference
is all usable space. Since that's not actually the case due to the
global metadata reservation, we should provide the full picture to the
user.

This patch adds an ioctl that exports the size of the global metadata
reservation so that btrfs filesystem df can report it.

	Signed-off-by: Jeff Mahoney <jeffm@suse.com>
	Signed-off-by: Josef Bacik <jbacik@fusionio.com>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit 01e219e8069516cdb98594d417b8bb8d906ed30d)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>