btrfs: expand btrfs_find_item() to include find_orphan_item functionality
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Kelley Nielsen <kelleynnn@gmail.com>
commit 3f870c28990015a1fd6c67807efcdb02a75b35e1

This is the third step in bootstrapping the btrfs_find_item interface.
The function find_orphan_item(), in orphan.c, is similar to the two
functions already replaced by the new interface. It uses two parameters,
which are already present in the interface, and is nearly identical to
the function brought in in the previous patch.

Replace the two calls to find_orphan_item() with calls to
btrfs_find_item(), with the defined objectid and type that was used
internally by find_orphan_item(), a null path, and a null key. Add a
test for a null path to btrfs_find_item, and if it passes, allocate and
free the path. Finally, remove find_orphan_item().

	Signed-off-by: Kelley Nielsen <kelleynnn@gmail.com>
	Signed-off-by: Josef Bacik <jbacik@fusionio.com>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit 3f870c28990015a1fd6c67807efcdb02a75b35e1)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>