rwsem: add rwsem_is_contended
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Josef Bacik <jbacik@fusionio.com>
commit 4a444b1f06d259ce938a47048840260f71a91c84

Btrfs needs a simple way to know if it needs to let go of it's read lock on a
rwsem.  Introduce rwsem_is_contended to check to see if there are any waiters on
this rwsem currently.  This is just a hueristic, it is meant to be light and not
100% accurate and called by somebody already holding on to the rwsem in either
read or write.  Thanks,

	Signed-off-by: Josef Bacik <jbacik@fusionio.com>
	Signed-off-by: Chris Mason <clm@fb.com>
	Acked-by: Ingo Molnar <mingo@kernel.org>
(cherry picked from commit 4a444b1f06d259ce938a47048840260f71a91c84)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>