Btrfs: only process as many file extents as there are refs
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Josef Bacik <jbacik@fb.com>
commit 7ef81ac86c8a44ab9f4e6e04e1f4c9ea53615b8a

The backref walking code will search down to the key it is looking for and then
proceed to walk _all_ of the extents on the file until it hits the end.  This is
suboptimal with large files, we only need to look for as many extents as we have
references for that inode.  I have a testcase that creates a randomly written 4
gig file and before this patch it took 6min 30sec to do the initial send, with
this patch it takes 2min 30sec to do the intial send.  Thanks,

	Signed-off-by: Josef Bacik <jbacik@fb.com>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit 7ef81ac86c8a44ab9f4e6e04e1f4c9ea53615b8a)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>