Btrfs: change the members' order of btrfs_space_info structure to reduce the cache miss
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Miao Xie <miaox@cn.fujitsu.com>
commit 26b47ff65bcdff8473b87680d8f876c66208087b

It is better that the position of the lock is close to the data which is
protected by it, because they may be in the same cache line, we will load
less cache lines when we access them. So we rearrange the members' position
of btrfs_space_info structure to make the lock be closer to the its data.

	Signed-off-by: Miao Xie <miaox@cn.fujitsu.com>
	Reviewed-by: David Sterba <dsterba@suse.cz>
	Signed-off-by: Josef Bacik <jbacik@fb.com>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit 26b47ff65bcdff8473b87680d8f876c66208087b)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>