Btrfs: unset DCACHE_DISCONNECTED when mounting default subvol
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Josef Bacik <jbacik@fb.com>
commit 3a0dfa6a12e4bb64a434426ecb17d4842092db5e

A user was running into errors from an NFS export of a subvolume that had a
default subvol set.  When we mount a default subvol we will use d_obtain_alias()
to find an existing dentry for the subvolume in the case that the root subvol
has already been mounted, or a dummy one is allocated in the case that the root
subvol has not already been mounted.  This allows us to connect the dentry later
on if we wander into the path.  However if we don't ever wander into the path we
will keep DCACHE_DISCONNECTED set for a long time, which angers NFS.  It doesn't
appear to cause any problems but it is annoying nonetheless, so simply unset
DCACHE_DISCONNECTED in the get_default_root case and switch btrfs_lookup() to
use d_materialise_unique() instead which will make everything play nicely
together and reconnect stuff if we wander into the defaul subvol path from a
different way.  With this patch I'm no longer getting the NFS errors when
exporting a volume that has been mounted with a default subvol set.  Thanks,

cc: bfields@fieldses.org
cc: ebiederm@xmission.com
	Signed-off-by: Josef Bacik <jbacik@fb.com>
	Acked-by: "Eric W. Biederman" <ebiederm@xmission.com>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit 3a0dfa6a12e4bb64a434426ecb17d4842092db5e)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>