Btrfs: don't mix the ordered extents of all files together during logging the inodes
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Miao Xie <miaox@cn.fujitsu.com>
commit 827463c49f29111efd22148d24c9ca44d648acfa

There was a problem in the old code:
If we failed to log the csum, we would free all the ordered extents in the log list
including those ordered extents that were logged successfully, it would make the
log committer not to wait for the completion of the ordered extents.

This patch doesn't insert the ordered extents that is about to be logged into
a global list, instead, we insert them into a local list. If we log the ordered
extents successfully, we splice them with the global list, or we will throw them
away, then do full sync. It can also reduce the lock contention and the traverse
time of list.

	Signed-off-by: Miao Xie <miaox@cn.fujitsu.com>
	Signed-off-by: Josef Bacik <jbacik@fb.com>
(cherry picked from commit 827463c49f29111efd22148d24c9ca44d648acfa)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>