Btrfs: just wait or commit our own log sub-transaction
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Miao Xie <miaox@cn.fujitsu.com>
commit d1433debe7f4346cf9fc0dafc71c3137d2a97bc4

We might commit the log sub-transaction which didn't contain the metadata we
logged. It was because we didn't record the log transid and just select
the current log sub-transaction to commit, but the right one might be
committed by the other task already. Actually, we needn't do anything
and it is safe that we go back directly in this case.

This patch improves the log sync by the above idea. We record the transid
of the log sub-transaction in which we log the metadata, and the transid
of the log sub-transaction we have committed. If the committed transid
is >= the transid we record when logging the metadata, we just go back.

	Signed-off-by: Miao Xie <miaox@cn.fujitsu.com>
	Signed-off-by: Josef Bacik <jbacik@fb.com>
(cherry picked from commit d1433debe7f4346cf9fc0dafc71c3137d2a97bc4)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>