Btrfs: fix preallocate vs double nocow write
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Miao Xie <miaox@cn.fujitsu.com>
commit 7b2b70851f862b68714f357d2926adbb6c574fdd

We can not release the reserved metadata space for the first write if we
find the write position is pre-allocated. Because the kernel might write
the data on the disk before we do the second write but after the can-nocow
check, if we release the space for the first write, we might fail to update
the metadata because of no space.

Fix this problem by end nocow write if there is dirty data in the range whose
space is pre-allocated.

	Signed-off-by: Miao Xie <miaox@cn.fujitsu.com>
	Signed-off-by: Josef Bacik <jbacik@fb.com>
(cherry picked from commit 7b2b70851f862b68714f357d2926adbb6c574fdd)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>