Btrfs: return EPERM when deleting a default subvolume
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Guangyu Sun <guangyu.sun@oracle.com>
commit 72de6b5393c15c5228074008bbdc47e92bf6d4f7

The error message is confusing:

 # btrfs sub delete /mnt/mysub/
 Delete subvolume '/mnt/mysub'
 ERROR: cannot delete '/mnt/mysub' - Directory not empty

The error message does not make sense to me: It's not about deleting a
directory but it's a subvolume, and it doesn't matter if the subvolume is
empty or not.

Maybe EPERM or is more appropriate in this case, combined with an explanatory
kernel log message. (e.g. "subvolume with ID 123 cannot be deleted because
it is configured as default subvolume.")

	Reported-by: Koen De Wit <koen.de.wit@oracle.com>
	Signed-off-by: Guangyu Sun <guangyu.sun@oracle.com>
	Reviewed-by: David Sterba <dsterba@suse.cz>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit 72de6b5393c15c5228074008bbdc47e92bf6d4f7)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>