Btrfs: do not reset last_snapshot after relocation
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Josef Bacik <jbacik@fb.com>
commit ba8b0289333a70f0d69923fd63350d0b0201b904

This was done to allow NO_COW to continue to be NO_COW after relocation but it
is not right.  When relocating we will convert blocks to FULL_BACKREF that we
relocate.  We can leave some of these full backref blocks behind if they are not
cow'ed out during the relocation, like if we fail the relocation with ENOSPC and
then just drop the reloc tree.  Then when we go to cow the block again we won't
lookup the extent flags because we won't think there has been a snapshot
recently which means we will do our normal ref drop thing instead of adding back
a tree ref and dropping the shared ref.  This will cause btrfs_free_extent to
blow up because it can't find the ref we are trying to free.  This was found
with my ref verifying tool.  Thanks,

	Signed-off-by: Josef Bacik <jbacik@fb.com>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit ba8b0289333a70f0d69923fd63350d0b0201b904)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>