Btrfs: take into account total references when doing backref lookup
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Josef Bacik <jbacik@fb.com>
commit 4485386853454f184235c8a973b29fa7fa522eb1

I added an optimization for large files where we would stop searching for
backrefs once we had looked at the number of references we currently had for
this extent.  This works great most of the time, but for snapshots that point to
this extent and has changes in the original root this assumption falls on it
face.  So keep track of any delayed ref mods made and add in the actual ref
count as reported by the extent item and use that to limit how far down an inode
we'll search for extents.  Thanks,

Reportedy-by: Hugo Mills <hugo@carfax.org.uk>
	Signed-off-by: Josef Bacik <jbacik@fb.com>
	Reported-by: Hugo Mills <hugo@carfax.org.uk>
	Tested-by: Hugo Mills <hugo@carfax.org.uk>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit 4485386853454f184235c8a973b29fa7fa522eb1)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>