Btrfs: use correct key when repeating search for extent item
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Filipe Manana <fdmanana@gmail.com>
commit 9ce49a0b4ff7f13961d8d106ffae959823d2e758

If skinny metadata is enabled and our first tree search fails to find a
skinny extent item, we may repeat a tree search for a "fat" extent item
(if the previous item in the leaf is not the "fat" extent we're looking
for). However we were not setting the new key's objectid to the right
value, as we previously used the same key variable to peek at the previous
item in the leaf, which has a different objectid. So just set the right
objectid to avoid modifying/deleting a wrong item if we repeat the tree
search.

	Signed-off-by: Filipe David Borba Manana <fdmanana@gmail.com>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit 9ce49a0b4ff7f13961d8d106ffae959823d2e758)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>