btrfs: fix inline compressed read err corruption
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Zach Brown <zab@redhat.com>
commit 166ae5a418756dacd2e8883c6f562c2d0ce2db2d

uncompress_inline() is dropping the error from btrfs_decompress() after
testing it and zeroing the page that was supposed to hold decompressed
data.  This can silently turn compressed inline data in to zeros if
decompression fails due to corrupt compressed data or memory allocation
failure.

I verified this by manually forcing the error from btrfs_decompress()
for a silly named copy of od:

	if (!strcmp(current->comm, "failod"))
		ret = -ENOMEM;

  # od -x /mnt/btrfs/dir/80 | head -1
  0000000 3031 3038 310a 2d30 6f70 6e69 0a74 3031
  # echo 3 > /proc/sys/vm/drop_caches
  # cp $(which od) /tmp/failod
  # /tmp/failod -x /mnt/btrfs/dir/80 | head -1
  0000000 0000 0000 0000 0000 0000 0000 0000 0000

The fix is to pass the error to its caller.  Which still has a BUG_ON().
So we fix that too.

There seems to be no reason for the zeroing of the page on the error
from btrfs_decompress() but not from the allocation error a few lines
above.  So the page zeroing is removed.

	Signed-off-by: Zach Brown <zab@redhat.com>
	Reviewed-by: David Sterba <dsterba@suse.cz>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit 166ae5a418756dacd2e8883c6f562c2d0ce2db2d)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>