Btrfs: send, don't error in the presence of subvols/snapshots
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Filipe Manana <fdmanana@gmail.com>
commit 1af56070e3ef9477dbc7eba3b9ad7446979c7974

If we are doing an incremental send and the base snapshot has a
directory with name X that doesn't exist anymore in the second
snapshot and a new subvolume/snapshot exists in the second snapshot
that has the same name as the directory (name X), the incremental
send would fail with -ENOENT error. This is because it attempts
to lookup for an inode with a number matching the objectid of a
root, which doesn't exist.

Steps to reproduce:

    mkfs.btrfs -f /dev/sdd
    mount /dev/sdd /mnt

    mkdir /mnt/testdir
    btrfs subvolume snapshot -r /mnt /mnt/mysnap1

    rmdir /mnt/testdir
    btrfs subvolume create /mnt/testdir
    btrfs subvolume snapshot -r /mnt /mnt/mysnap2

    btrfs send -p /mnt/mysnap1 /mnt/mysnap2 -f /tmp/send.data

A test case for xfstests follows.

	Reported-by: Robert White <rwhite@pobox.com>
	Signed-off-by: Filipe David Borba Manana <fdmanana@gmail.com>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit 1af56070e3ef9477dbc7eba3b9ad7446979c7974)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>