Btrfs: make fsync work after cloning into a file
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Filipe Manana <fdmanana@gmail.com>
commit 7ffbb598a059b73487909619d73150f99b50337a

When cloning into a file, we were correctly replacing the extent
items in the target range and removing the extent maps. However
we weren't replacing the extent maps with new ones that point to
the new extents - as a consequence, an incremental fsync (when the
inode doesn't have the full sync flag) was a NOOP, since it relies
on the existence of extent maps in the modified list of the inode's
extent map tree, which was empty. Therefore add new extent maps to
reflect the target clone range.

A test case for xfstests follows.

	Signed-off-by: Filipe David Borba Manana <fdmanana@gmail.com>
	Signed-off-by: Chris Mason <clm@fb.com>
(cherry picked from commit 7ffbb598a059b73487909619d73150f99b50337a)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>