virtio: rng: add derating factor for use by hwrng core
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
Rebuild_CHGLOG: - [hwrng] virtio-rng: add derating factor for use by hwrng core (Amit Shah) [1129195]
Rebuild_FUZZ: 97.20%
commit-author Amit Shah <amit.shah@redhat.com>
commit 34679ec7a0c45da8161507e1f2e1f72749dfd85c

The khwrngd thread is started when a hwrng device of sufficient
quality is registered.  The virtio-rng device is backed by the
hypervisor, and we trust the hypervisor to provide real entropy.

A malicious or badly-implemented hypervisor is a scenario that's
irrelevant -- such a setup is bound to cause all sorts of badness, and a
compromised hwrng is the least of the user's worries.

Given this, we might as well assume that the quality of randomness we
receive is perfectly trustworthy.  Hence, we use 100% for the factor,
indicating maximum confidence in the source.

	Signed-off-by: Amit Shah <amit.shah@redhat.com>
	Reviewed-by: H. Peter Anvin <hpa@linux.intel.com>
	Reviewed-by: Amos Kong <akong@redhat.com>
	Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>
(cherry picked from commit 34679ec7a0c45da8161507e1f2e1f72749dfd85c)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>