igb: enable VLAN stripping for VMs with i350
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Stefan Assmann <sassmann@kpanic.de>
commit dc1edc67fe2da3d87b56cdffd5ef55c3a9af252c

For i350 VLAN stripping for VMs is not enabled in the VMOLR register but in
the DVMOLR register. Making the changes accordingly. It's not necessary to
unset the E1000_VMOLR_STRVLAN bit on i350 as the hardware will simply ignore
it.

Without this change if a VLAN is configured for a VF assigned to a guest
via (i.e.)
ip link set p1p1 vf 0 vlan 10
the VLAN tag will not be stripped from packets going into the VM. Which they
should be because the VM itself is not aware of the VLAN at all.

	Signed-off-by: Stefan Assmann <sassmann@kpanic.de>
	Tested-by: Sibai Li <sibai.li@intel.com>
	Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
(cherry picked from commit dc1edc67fe2da3d87b56cdffd5ef55c3a9af252c)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>