cifs: fix potential races in cifs_revalidate_mapping
jira LE-1907
Rebuild_History Non-Buildable kernel-3.10.0-229.el7
commit-author Jeff Layton <jlayton@poochiereds.net>
commit 4f73c7d342d57d065bdbc0995cb56d8d1701b0c0

The handling of the CIFS_INO_INVALID_MAPPING flag is racy. It's possible
for two tasks to attempt to revalidate the mapping at the same time. The
first sees that CIFS_INO_INVALID_MAPPING is set. It clears the flag and
then calls invalidate_inode_pages2 to start shooting down the pagecache.

While that's going on, another task checks the flag and sees that it's
clear. It then ends up trusting the pagecache to satisfy a read when it
shouldn't.

Fix this by adding a bitlock to ensure that the clearing of the flag is
atomic with respect to the actual cache invalidation. Also, move the
other existing users of cifs_invalidate_mapping to use a new
cifs_zap_mapping() function that just sets the INVALID_MAPPING bit and
then uses the standard codepath to handle the invalidation.

	Signed-off-by: Jeff Layton <jlayton@poochiereds.net>
	Signed-off-by: Steve French <smfrench@gmail.com>
(cherry picked from commit 4f73c7d342d57d065bdbc0995cb56d8d1701b0c0)
	Signed-off-by: Jonathan Maple <jmaple@ciq.com>